Renaming FeatureToggle Model to Database to match naming.
