+++ tests/src/Unit/Toggle/FeatureToggleTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace FeatureToggle\Tests\Unit\Toggle;
 
 use FeatureToggle\Tests\TestCase;
-use FeatureToggle\Toggle\FeatureToggle;
+use FeatureToggle\Toggle\Database;
 use FeatureToggle\Tests\Traits\TestToggle;
 use FeatureToggle\Contracts\Toggle as ToggleContract;
 
@@ -16,10 +16,10 @@ class FeatureToggleTest extends TestCase
     /**
      * @param  string  $name
      * @param  mixed  $is_active
-     * @return ToggleContract|FeatureToggle
+     * @return ToggleContract|Database
      */
     protected function getInstance(string $name, $is_active): ToggleContract
     {
-        return new FeatureToggle(compact('name', 'is_active'));
+        return new Database(compact('name', 'is_active'));
     }
 }
